UXENG-2188 Use symlinks for plugins
